# openapi/apexrouter-v1.yaml
openapi: 3.1.0
info:
  title: ApexRouter-RS
  version: 0.1.0
  summary: One local router in front of every OpenAI-compatible endpoint you can reach.
  description: >-
    The machine-readable form of ARCHITECTURE.md section 6. Two listeners: the proxy plane
    (the OpenAI/Anthropic drop-in surface) and the control plane (rig, fit, routes, endpoints,
    providers, vast, hf, jobs). Every operation carries an `x-listener` extension naming which
    listener serves it; `/health` is the only path served by both. Prose lives in docs/API.md;
    routing semantics live in docs/ROUTING.md. This file is diffed against the axum route table
    by crates/apexrouter-server/tests/openapi_routes.rs on every CI run.
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
  contact:
    name: ApexRouter-RS
    url: https://github.com/buckster123/ApexRouter-RS
servers:
  - url: http://127.0.0.1:8888
    description: 'proxy listener. `http://127.0.0.1:8888/v1` is equally valid: a doubled leading /v1 is collapsed.'
  - url: http://127.0.0.1:2739
    description: control listener.
...
security:
  - bearerAuth: []
  - tokenHeader: []
  - tokenQuery: []
paths:
...
  /v1/hf/models/{repo}/files:
    get:
      operationId: hfRepoFiles
      x-listener: [control]
      tags: [hf]
      x-axum-route: /v1/hf/models/{*rest}
      summary: Authoritative per-file sizes from paths-info, grouped by quant.
      description: >-
        `repo` contains a slash, so axum registers this as the wildcard `/v1/hf/models/{*rest}` and
        the handler splits the trailing `/files` itself.
      parameters:
        - name: repo
          in: path
          required: true
          description: 'e.g. unsloth/Qwen3-Coder-30B-A3B-Instruct-GGUF'
          schema: { type: string }
      responses:
        '200':
          description: The same grouping the UI shows, so what downloads is what was clicked.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/HfFileGroup' }
        '404': { $ref: '#/components/responses/ControlError' }
  /v1/hf/downloads:
    get:
      operationId: listHfDownloads
      x-listener: [control]
      tags: [hf]
      summary: Download jobs.
      responses:
        '200':
          description: Job rows.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/JobRecord' }
    post:
      operationId: startHfDownload
      x-listener: [control]
      tags: [hf]
      summary: Download named files, or every shard of a quant.
      parameters: [{ $ref: '#/components/parameters/NoWait' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HfDownloadRequest' }
      responses:
        '202':
          description: Accepted; poll the job.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobRecord' }
        '400': { $ref: '#/components/responses/ControlError' }
...
  /v1/compare:
    post:
      operationId: compare
      x-listener: [control]
      tags: [observability]
      summary: The same prompt against N aliases, at the same time.
      description: Blocking by default, because the answer IS the comparison.
      parameters: [{ $ref: '#/components/parameters/NoWait' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompareRequest' }
      responses:
        '200':
          description: One row per alias.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CompareRow' }
        '202':
          description: Accepted; the job's result is the same array.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobRecord' }
...
  /v1/jobs/{id}:
    get:
      operationId: getJob
      x-listener: [control]
      tags: [observability]
      summary: One job.
      parameters: [{ $ref: '#/components/parameters/JobId' }]
      responses:
        '200':
          description: The job.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobRecord' }
        '404': { $ref: '#/components/responses/ControlError' }
...
  /v1/migrate:
    post:
      operationId: migrate
      x-listener: [control]
      tags: [observability]
      summary: Import ~/.vastai-gguf. dry_run writes nothing.
      description: >-
        --apply imports providers as credential REFERENCES; the real Together key is not copied.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MigrateRequest' }
      responses:
        '200':
          description: A MigrationPlan under dry_run, a MigrationReport otherwise.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MigrationPlan'
                  - $ref: '#/components/schemas/MigrationReport'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Authorization: Bearer <token>. Scopes read|write|admin are derived from (path, method).'
    tokenHeader:
      type: apiKey
      in: header
      name: X-ApexRouter-Token
      description: The same token, as a header.
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: The same token, as a query parameter. Never recorded in a tracing span.
  parameters:
...
    NoWait:
      name: no_wait
      in: query
      required: false
      description: Return a JobRecord immediately instead of waiting.
      schema: { type: boolean }
...
    JobId:
      name: id
      in: path
      required: true
      description: A ULID.
      schema: { type: string }
...
  responses:
...
    ControlError:
      description: A control-plane ErrorEnvelope. Clients branch on error.kind, never on prose.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    BackendId:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
      description: A validated slug. An empty or malformed id is not constructible.
    Alias:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
      description: The string a client puts in "model".
...
    TokenCount:
      type: object
      required: [kind, n]
      properties:
        kind: { type: string, enum: [reported, estimated] }
        n: { type: integer, format: int64 }

    # ---- errors -----------------------------------------------------------------------
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
    ErrorBody:
      type: object
      required: [kind, message]
      properties:
        kind:
          type: string
          description: 'The discriminator, so no exit codes had to be invented.'
          examples: [not_found, invalid_request, conflict, port_in_use, unauthorized, forbidden,
                     upstream_unavailable, approval_required, internal]
        message: { type: string }
        param: { type: [string, 'null'] }
        code: { type: [string, 'null'] }
...
    JobRecord:
      type: object
      required: [id, kind, state, started_unix]
      properties:
        id: { type: string }
        kind: { type: string, examples: [hf.download, vast.rent, endpoint.start, compare, swap] }
        state:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
          description: Never Pending forever; even a JoinError from a panic flips it to failed.
        pct: { type: [number, 'null'] }
        message: { type: [string, 'null'] }
        started_unix: { type: integer, format: int64 }
        finished_unix: { type: [integer, 'null'], format: int64 }
        result: { description: 'The protocol type the operation produced, as JSON.' }
        error: { type: [string, 'null'] }

    # ---- control plane: rig and fit -----------------------------------------------------
...
    HfFile:
      type: object
      required: [rfilename, is_mmproj]
      properties:
        rfilename: { type: string, description: Path within the repo. }
        size: { type: [integer, 'null'], format: int64, description: 'Bytes, from paths-info.' }
        quant: { type: [string, 'null'] }
        is_mmproj:
          type: boolean
          description: 'Matched as a filename TOKEN, not a path substring, so a directory named vocab-x hides nothing.'
        shard_of:
          type: ['array', 'null']
          description: '(index, total) parsed from -00001-of-000NN.'
          items: { type: integer, format: int64 }
    HfFileGroup:
      type: object
      description: 'Files grouped into one downloadable unit: a quant plus its shards plus its projector.'
      required: [label, total_bytes]
      properties:
        label: { type: string }
        quant: { type: [string, 'null'] }
        total_bytes:
          type: integer
          format: int64
          description: Sum of every file's size — the number the fit solver and the disk check use.
        files: { type: array, items: { $ref: '#/components/schemas/HfFile' } }
        mmproj: { type: array, items: { $ref: '#/components/schemas/HfFile' } }
    HfDownloadRequest:
      type: object
      required: [repo]
      properties:
        repo: { type: string }
        files:
          type: array
          items: { type: string }
          description: Exact repo-relative paths. Wins over `quant` when both are given.
        quant: { type: [string, 'null'] }
        dest: { type: [string, 'null'], description: 'null uses [hf] download_dir.' }

    # ---- control plane: observability --------------------------------------------------------
...
    CompareRequest:
      type: object
      required: [prompt]
      properties:
        aliases: { type: array, items: { type: string } }
        prompt: { type: string }
        max_tokens: { type: [integer, 'null'], format: int64 }
    CompareRow:
      type: object
      properties:
        alias: { $ref: '#/components/schemas/Alias' }
        backend: { oneOf: [{ $ref: '#/components/schemas/BackendId' }, { type: 'null' }] }
        model: { type: string }
        ok: { type: boolean }
        ms: { type: integer, format: int64 }
        ttft_ms: { type: [integer, 'null'], format: int64 }
        tok_per_s: { type: [number, 'null'] }
        prompt_tokens:
          description: 'The REAL number from the response, never word_count × 1.3.'
          oneOf: [{ $ref: '#/components/schemas/TokenCount' }, { type: 'null' }]
        text: { type: [string, 'null'] }
      additionalProperties: true
...
    MigrateRequest:
      type: object
      required: [dry_run]
      properties:
        dry_run: { type: boolean, description: 'true returns a MigrationPlan and writes NOTHING.' }
    MigrationItem:
      type: object
      properties:
        what: { type: string, examples: [recipe, usage row, known_fork, provider, instance] }
        from: { type: string }
        action: { type: string, enum: [import, skip, warn] }
        detail: { type: string, description: 'The reason, per row.' }
    MigrationPlan:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/MigrationItem' } }
        source_paths: { type: array, items: { type: string } }
    MigrationReport:
      type: object
      properties:
        imported: { type: integer, format: int64 }
        skipped: { type: integer, format: int64 }
        warnings: { type: array, items: { type: string } }